[#361] Rename template and sample application class to MainApplication
